Switched placeholder data over to an API service, reorganized files, refactored code
